+++ CHANGELOG.md
@@ -20,6 +20,9 @@
 * Added `eval-jexl` command to nimbus-cli for evaluating JEXL targeting expressions against the app context. Useful for testing and debugging targeting expressions on iOS and Android.
 ([#7160](https://github.com/mozilla/application-services/pull/7160))
 
+### Ads-Client
+* Adds new Kotlin `AdsClientTelemetry.kt` wrapper for Glean callbacks.
+
 [Full Changelog](In progress)
 
 # v148.0 (_2026-01-12_)